Fix table interpolation in xspec module
xspec called tabint via the fortan interface "provided" by Xspec.
The new heasoft version v6.27.1 properly declares tabint as extern "C"
so it is better to call tabint directly rather than using the string +
long hack.

More severe, the call to tabint_ was plain wrong and caused segfaults
when evaluating table models. It is surprising, that this did not show
up earlier. However, it is easy to understand why it only happens
occasionally. Some of the arguments are ment to be int and were given
as int*. Since those are only used to initialize the valarrays it works
as long as they turn into positive integers. If they are negative
-> segfault.