Fix memory leak on exit with xspec module
if putenv is called on a variable that exists it will free the old memory
location and allocate a new one (not sure if this is true for the static
environment memory location). If the xspec-module sets HEADAS and
later xspec itself set HEADAS again the pointer given to putenv in the
xspec-module is free'd. Calling free on this pointer in this case is not
a good idea. So we check if the pointer set by the module is still the
same as the environemnt variable and only if free that pointer.